ejam coding task
##Given files - price_train.csv and price_test.csv

1]] Notebook with the following components and partial scores for each component: 
a. EDA- documented in the notebook, with graphs describing correlations between variables, potential predictors, initial analyses on the data, and feature engineering (if any)
b. Data Engineering - documented in the notebook, in a few sentences describing the ETL process and any data engineering that was performed 
c. Initial Modelling - a few models run on smaller folds of the dataset, with explanations for why each model was experimented with 
d. Model Selection - analyses around the output from each of the models initially selected, and justification for selecting one model over the others you had initially contemplated 

2]] CSV file of listing prices for the test set: a. Final Predictions - each listing from the data set and the model-predicted price (2 columns: id, price)




***Disclaimer***
This is my personal coding task for the company eJam. The deadline was within three days, so I managed to make a better model as possible. This was part of my interview process for the same company. It may not be possible that others have the same opinion but this is my own thoughts before uploading it publically via GitHub profile. Thanks for understanding.